requirements.txt
numpy
pyserial
setuptools
